feat: extract ProtectedRoute component
