+++ src/hooks/useSaveRedirect.tsx
@@ -1,0 +1,20 @@
+import { useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
+import { useAuth } from '../contexts/AuthContext';
+
+/**
+ * Persists the user's intended destination when they hit a protected route.
+ * Stores the current pathname in sessionStorage if unauthenticated.
+ */
+const useSaveRedirect = () => {
+  const { user, loading, isDemo } = useAuth();
+  const location = useLocation();
+
+  useEffect(() => {
+    if (!user && !loading && !isDemo) {
+      sessionStorage.setItem('redirectUrl', location.pathname);
+    }
+  }, [user, loading, isDemo, location]);
+};
+
+export default useSaveRedirect;
